requirements.txt
tensorflow-gpu
tqdm
